fix: resolve all compilation errors and database crash on launch
Compilation fixes:
- Downgrade kotlinx-serialization 1.7.1 → 1.6.3 (Kotlin 1.9.24 compat)
- Fix CornerRadius import (geometry not graphics)
- Fix brace mismatch in WizardViewModel.kt
- Fix WizardScreen.kt smart cast and unresolved references
- Fix DimensionPreviewCanvas.kt Float/Double type mismatches
- Fix FloorTypeSelector.kt, WallMaterialSelector.kt, HouseDimensionsStep.kt refs
- Fix Type.kt font references
- Add font resources (Outfit family)

Database crash fix:
- Remove createFromAsset('seed/poultry.db') — pre-packaged DB had invalid schema
- Room now creates database from scratch with correct schema v3
- Seed default breed profiles (Cobb 500, Ross 308) via RoomDatabase.Callback
- Add .gitignore for SQLite temp files (*.db-shm, *.db-wal)

+++ app/src/main/java/com/poultry/broiler/presentation/wizard/components/DimensionPreviewCanvas.kt
@@ -87,7 +87,7 @@ fun DimensionPreviewCanvas(
         val usableHeight = canvasHeightPx - 2 * paddingPx
 
         val rect: Rect = if (length != null && width != null && length > 0.0 && width > 0.0) {
-            val aspect = length / width
+            val aspect = (length / width).toFloat()
             var rectWidth = usableWidth
             var rectHeight = usableWidth / aspect
             if (rectHeight > usableHeight) {
